refactor: delete socket io namespaces
