+++ src/controllers/video/videoController.ts
@@ -1,0 +1,1 @@
+export interface VideoController {}
