+++ src/view/commands/connection.ts
@@ -1,24 +1,24 @@
-import { Namespace, Socket } from 'socket.io'
-import { commandListener, commandListenerWithVerification } from '../../utils'
+import { Server, Socket } from 'socket.io'
 import { disconnectionCommand } from './disconnect'
 import { userTokenCommand } from './userToken'
-import { ChatController } from '../../../controllers/chat/chatController'
+import { ChatController } from '../../controllers/chat/chatController'
+import { commandListener, commandListenerWithVerification } from '../utils'
 
 /**
  * On connection command.
  * After the client successfully creates a connection with the chat namespace, it enables the client to further:
  * (1) send the token as first initialization step;
  * (2) disconnect from the namespace.
- * @param chatNamespace
+ * @param io
  * @param chatController
  */
-export function connectionCommand(chatNamespace: Namespace, chatController: ChatController) {
+export function connectionCommand(io: Server, chatController: ChatController) {
   commandListenerWithVerification(
-    chatNamespace,
+    io,
     'connection',
     () => true,
     (socket: Socket) => {
-      commandListener(socket, 'userToken', userTokenCommand(chatNamespace, socket, chatController))
+      commandListener(socket, 'userToken', userTokenCommand(io, socket, chatController))
       commandListener(socket, 'disconnect', disconnectionCommand(socket))
     }
   )
