+++ src/view/commands/joinRoom.ts
@@ -1,10 +1,10 @@
-import { Namespace, Socket } from 'socket.io'
-import { ChatUpdate } from '../../../model/message'
-import { ChatController } from '../../../controllers/chat/chatController'
-import { commandListener, reaction } from '../../utils'
+import { Server, Socket } from 'socket.io'
+import { ChatUpdate } from '../../model/message'
+import { ChatController } from '../../controllers/chat/chatController'
 import { leaveRoomCommand } from './leaveRoom'
+import { SerializerImpl } from '../../model/presentation/serialization/messageSerializer'
 import { sendMessageCommand } from './sendMessage'
-import { SerializerImpl } from '../../../model/presentation/serialization/messageSerializer'
+import { commandListener, reaction } from '../utils'
 
 /**
  * Join Command.
@@ -12,14 +12,14 @@ import { SerializerImpl } from '../../../model/presentation/serialization/messag
  * After the client successfully joins a room, it enables the client to further:
  * (1) send a message to the room;
  * (2) leave the room.
- * @param chatNamespace
+ * @param io
  * @param socket
  * @param token
  * @param chatController
  * @returns
  */
 export function joinCommand(
-  chatNamespace: Namespace,
+  io: Server,
   socket: Socket,
   token: string,
   chatController: ChatController
@@ -32,24 +32,19 @@ export function joinCommand(
         reaction(
           chatController.joinUserToRoom(token, room),
           (chatUpdate: ChatUpdate) => {
-            chatNamespace
-              .to(room)
-              .emit(
-                'notificationMessage',
-                new SerializerImpl().serialize(chatUpdate.notificationMessage)
-              )
+            io.to(room).emit(
+              'notificationMessage',
+              new SerializerImpl().serialize(chatUpdate.notificationMessage)
+            )
             socket.join(room)
             socket.emit('chatUpdate', new SerializerImpl().serialize(chatUpdate.messages))
             commandListener(
               socket,
               'leaveRoom',
-              leaveRoomCommand(chatNamespace, socket, room, token, chatController)
+              leaveRoomCommand(io, socket, room, token, chatController)
             )
-            commandListener(
-              socket,
-              'sendMessage',
-              sendMessageCommand(chatNamespace, token, room, chatController)
-            )
+            defineChatCommands(io, socket, token, room, chatController)
+            // Define video commands
           },
           ack
         )
@@ -57,4 +52,14 @@ export function joinCommand(
       ack
     )
   }
+}
+
+function defineChatCommands(
+  io: Server,
+  socket: Socket,
+  token: string,
+  room: string,
+  chatController: ChatController
+) {
+  commandListener(socket, 'sendMessage', sendMessageCommand(io, token, room, chatController))
 }
