+++ src/view/commands/leaveRoom.ts
@@ -1,20 +1,20 @@
-import { Namespace, Socket } from 'socket.io'
-import { NotificationMessage } from '../../../model/message'
-import { ChatController } from '../../../controllers/chat/chatController'
-import { reaction } from '../../utils'
-import { SerializerImpl } from '../../../model/presentation/serialization/messageSerializer'
+import { Server, Socket } from 'socket.io'
+import { NotificationMessage } from '../../model/message'
+import { ChatController } from '../../controllers/chat/chatController'
+import { reaction } from '../utils'
+import { SerializerImpl } from '../../model/presentation/serialization/messageSerializer'
 
 /**
  * Leave command.
  * Using the token and socket, leaves the user to the specified room.
- * @param chatNamespace
+ * @param io
  * @param socket
  * @param room
  * @param chatController
  * @returns
  */
 export function leaveRoomCommand(
-  chatNamespace: Namespace,
+  io: Server,
   socket: Socket,
   room: string,
   token: string,
@@ -26,9 +26,7 @@ export function leaveRoomCommand(
       (notificationMessage: NotificationMessage) => {
         socket.leave(room /*, token*/)
         socket.disconnect()
-        chatNamespace
-          .to(room)
-          .emit('notificationMessage', new SerializerImpl().serialize(notificationMessage))
+        io.to(room).emit('notificationMessage', new SerializerImpl().serialize(notificationMessage))
       }
     )
   }
