+++ src/view/defineNamespaces.ts
@@ -1,14 +1,7 @@
 import { Server } from 'socket.io'
-import { ChatNamespace } from './chat/chatNamespace'
-import { SessionNamespace } from './sessionNamespace'
-import { VideoNamespace } from './commands/videoCommands'
+import { connectionCommand } from './commands/connection'
+import { ChatController } from '../controllers/chat/chatController'
 
 export async function registerCommands(io: Server) {
-  // Register Chat Commands
-  const chatNamespace: SessionNamespace = new ChatNamespace()
-  chatNamespace.registerCommands(io.of('/chat'))
-
-  //  Register Video Commands
-  const videoNamespace: VideoNamespace = new VideoNamespace()
-  videoNamespace.registerCommands(io.of('/video'))
+  connectionCommand(io, new ChatController())
 }
