--- src/view/sessionNamespace.ts
@@ -1,5 +1,0 @@
-import { Namespace } from 'socket.io'
-
-export interface SessionNamespace {
-  registerCommands(namespace: Namespace): void
-}
